refactored styles into smaller files, added smaller components (buttons, comment, rating)
