+++ client/src/styles/comment.style.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const CommentStyle = styled.div`
+  width: auto;
+  max-width: 600px;
+`;
