+++ client/src/styles/main.style.js
@@ -1,0 +1,31 @@
+import styled from 'styled-components';
+
+export const MainStyle = styled.div`
+  box-sizing: border-box;
+  color: rgb(60, 59, 55);
+  display: block;
+  font-family: "sf pro text", -apple-system, system-ui, Roboto, "segoe ui", Helvetica, Arial, sans-serif, "apple color emoji", "segoe ui emoji", "segoe ui symbol";
+  font-size: 14px;
+  font-weight: 400;
+  float: left;
+  letter-spacing: -0.2px
+  line-height: 22.4px;
+  min-height 701px;
+  width: auto;
+  min-width: 600px;
+  max-width: 1084px;
+  padding-bottom: 0px;
+  padding-left: 32px;
+  padding-top: 32px;
+  text-size-adjust: 100%;
+  -webkit-font-smoothing: antialiased;
+  -webkit-tap-highlight-color: rgba(0, 0, 0, 0);
+`;
+
+export const Title = styled.h2`
+  font-size: 24px;
+  font-weight: 700;
+  line-height: 28.8px;
+  margin-block-end: 16px;
+  margin-top: 0;
+`;
