+++ client/src/styles/rating.style.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+export const RatingStyle = styled.span`
+  margin: 8px 0;
+  display: flex;
+  align-items: center;
+  height: 20px;
+`;
+
+export const Moment = styled.span`
+  color: #73726c;
+  margin-left: 8px;
+`;
