+++ client/src/styles/reviewList.style.js
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+export const ReviewListStyle = styled.div`
+  margin-top: 12px;
+`;
