+++ client/src/styles/searchMessage.style.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const Message = styled.div`
+  font-size: 16px;
+  margin: 16px;
+  line-height: 22.4px;
+`;
