+++ client/src/styles/stars.style.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+export const StarStyle = styled.svg`
+  width: 16px;
+  height: 16px;
+  margin: 2px;
+`;
+
+export const StarsStyle = styled.div`
+  width: 100px;
+  height: 20px;
+`;
